docs: add further snippets with new infrastructure (#3297)
