+++ .changeset/lazy-years-share.md
@@ -1,0 +1,5 @@
+---
+"fuels": patch
+---
+
+docs: add further snippets with new infrastructure
