+++ .prettierignore
@@ -14,6 +14,8 @@ apps/demo-typegen/src/script-types
 apps/demo-typegen/src/predicate-types
 apps/docs/.vitepress/cache/
 apps/create-fuels-counter-guide
+apps/docs-snippets/.fuels
+apps/docs-snippets2/.fuels
 
 packages/fuels/src/cli/commands/deploy/proxy
 packages/fuels/test/fixtures/project
