+++ apps/docs-snippets/.gitignore
@@ -1,0 +1,1 @@
+.fuels/
