+++ apps/docs-snippets2/.gitignore
@@ -1,2 +1,3 @@
 src/typegend
 src/**/*.test.ts
+.fuels/
