+++ apps/docs-snippets2/package.json
@@ -7,7 +7,7 @@
     "build": "run-s wrap:snippets build:forc",
     "build:forc": "pnpm fuels build",
     "wrap:snippets": "tsx scripts/wrap-snippets.ts",
-    "pretest": "pnpm build",
+    "pretest": "./scripts/pretest.sh",
     "test": "cd ../.. && pnpm run test:filter apps/docs-snippets2"
   },
   "dependencies": {
