+++ apps/docs-snippets2/scripts/pretest.sh
@@ -1,5 +1,5 @@
 # Check if node is already running at port 4000, if not start it
-# TODO: This is a temporary solution to avoid conflicts with the test node in docs-snippets2
+# TODO: This is a temporary solution to avoid conflicts with the test node in docs-snippets
 if ! lsof -t -i:4000 > /dev/null; then
   pnpm fuels node > /dev/null 2>&1 &
 fi
