+++ apps/docs-snippets2/scripts/wrap-snippets.ts
@@ -12,6 +12,24 @@ const wrapperFnContents = readFileSync(wrapperFnFilepath, 'utf-8');
 export const wrapSnippet = (filepath: string) => {
   const snippetContents = readFileSync(filepath, 'utf8');
 
+  /**
+   * Test environment
+   */
+  let testEnvironments = '';
+
+  // Check if the file contains 'node' or 'browser' groups at the top of the file
+  const fileContents = readFileSync(filepath, 'utf8');
+  const hasNodeComment = fileContents.includes('@group node');
+  const hasBrowserComment = fileContents.includes('@group browser');
+
+  if (hasNodeComment && !hasBrowserComment) {
+    testEnvironments = '/**\n * @group node\n */';
+  } else if (hasBrowserComment && !hasNodeComment) {
+    testEnvironments = '/**\n * @group browser\n */';
+  } else {
+    testEnvironments = '/**\n * @group node\n * @group browser\n */';
+  }
+
   /*
     Filter all imports from file.
   */
@@ -74,6 +92,7 @@ export const wrapSnippet = (filepath: string) => {
   const wrappedSnippet =
     // eslintDisableRule +
     wrapperFnContents
+      .replace('// %TEST_ENVIRONMENT%', testEnvironments)
       .replace('// %IMPORTS%', imports)
       .replace('%NAME%', basename(filepath))
       .replace('// %SNIPPET%', indented)
