+++ apps/docs-snippets2/src/types/enums/using-an-invalid-enum-value.ts
@@ -1,10 +1,10 @@
 // #region snippet-1
 // Valid values: 'Void', 'Pending', 'Completed'
-const emumParam = 'NotStateEnumValue';
+const invalidEnumValue = 'NotStateEnumValue';
 
 try {
   // @ts-expect-error NotStateEnumValue is not a valid value
-  await contract.functions.echo_state_error_enum(emumParam).get();
+  await contract.functions.echo_state_error_enum(invalidEnumValue).get();
 } catch (error) {
   console.log('error', error);
 }
