+++ apps/docs-snippets2/sway/Forc.toml
@@ -2,6 +2,8 @@
 members = [
   "bytecode-input",
   "counter",
+  "echo-values",
+  "script-transfer-to-contract",
   "echo-asset-id",
   "echo-bytes",
   "echo-employee-data-vector",
@@ -9,7 +11,6 @@ members = [
   "echo-evm-address",
   "echo-raw-slice",
   "echo-std-string",
-  "echo-values",
   "employee-data",
   "input-output-types",
   "script-sum",
