+++ apps/docs-snippets2/sway/counter/src/main.sw
@@ -6,6 +6,9 @@ abi Counter {
 
     #[storage(write, read)]
     fn increment_count(amount: u64) -> u64;
+
+    #[storage(write, read)]
+    fn decrement_count(amount: u64) -> u64;
 }
 
 storage {
@@ -22,6 +25,13 @@ impl Counter for Contract {
     fn increment_count(amount: u64) -> u64 {
         let current = storage.counter.read();
         storage.counter.write(current + amount);
+        storage.counter.read()
+    }
+
+    #[storage(write, read)]
+    fn decrement_count(amount: u64) -> u64 {
+        let current = storage.counter.read();
+        storage.counter.write(current - amount);
         storage.counter.read()
     }
 }
