+++ apps/docs-snippets2/sway/script-transfer-to-contract/Forc.toml
@@ -1,0 +1,7 @@
+[project]
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+name = "script-transfer-to-contract"
+
+[dependencies]
