+++ apps/docs/src/guide/cookbook/custom-transactions.md
@@ -4,8 +4,14 @@ There may be scenarios where you need to build out transactions that involve mul
 
 Consider the following script that transfers multiple assets to a contract:
 
-<<< @/../../docs-snippets/test/fixtures/forc-projects/script-transfer-to-contract/src/main.sw#custom-transactions-1{ts:line-numbers}
+<<< @/../../docs-snippets2/sway/script-transfer-to-contract/src/main.sw#custom-transactions-1{rust:line-numbers}
 
 This script can be executed by creating a [`ScriptTransactionRequest`](../../api/Account/ScriptTransactionRequest.md), appending the resource and contract inputs/outputs and then sending the transaction, as follows:
 
-<<< @/../../docs-snippets/src/guide/scripts/script-custom-transaction.test.ts#custom-transactions-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/scripts/script-custom-transaction.ts#custom-transactions-2{ts:line-numbers}
+
+## Full Example
+
+For a full example, see below:
+
+<<< @/../../docs-snippets2/src/scripts/script-custom-transaction.ts#full{ts:line-numbers}
