+++ apps/docs/src/guide/encoding/encode-and-decode.md
@@ -11,7 +11,7 @@ The `Interface` class requires you to pass the [ABI](https://docs.fuel.network/d
 
 Imagine we are working with the following script that returns the sum of two `u32` integers:
 
-<<< @/../../docs-snippets/test/fixtures/forc-projects/sum-script/src/main.sw#encode-and-decode-1{rust:line-numbers}
+<<< @/../../docs-snippets2/sway/script-sum/src/main.sw#encode-and-decode-1{rust:line-numbers}
 
 When you build this script, using:
 
@@ -27,16 +27,22 @@ Now, let's prepare some data to pass to the `main` function to retrieve the comb
 
 First, let's prepare the transaction:
 
-<<< @/../../docs-snippets/src/guide/encoding/encode-and-decode.test.ts#encode-and-decode-3{ts:line-numbers}
+<<< @/../../docs-snippets2/src/encoding/encode-and-decode.ts#encode-and-decode-3{ts:line-numbers}
 
 Now, we can encode the script data to use in the transaction:
 
-<<< @/../../docs-snippets/src/guide/encoding/encode-and-decode.test.ts#encode-and-decode-4{ts:line-numbers}
+<<< @/../../docs-snippets2/src/encoding/encode-and-decode.ts#encode-and-decode-4{ts:line-numbers}
 
 Finally, we can decode the result:
 
-<<< @/../../docs-snippets/src/guide/encoding/encode-and-decode.test.ts#encode-and-decode-5{ts:line-numbers}
+<<< @/../../docs-snippets2/src/encoding/encode-and-decode.ts#encode-and-decode-5{ts:line-numbers}
 
 A similar approach can be taken with [Predicates](../predicates/index.md); however, you must set the encoded values to the `predicateData` property.
 
 [Contracts](../contracts/index.md) require more care. Although you can utilize the `scriptData` property, the arguments must be encoded as part of the [contract call script](https://docs.fuel.network/docs/sway/sway-program-types/smart_contracts/#calling-a-smart-contract-from-a-script). Therefore, it is recommended to use a `FunctionInvocationScope` when working with contracts which will be instantiated for you when [submitting a contract function](../contracts/methods.md), and therefore handles all the encoding.
+
+## Full Example
+
+Here is the full example of the encoding and decoding methods:
+
+<<< @/../../docs-snippets2/src/encoding/encode-and-decode.ts#full{ts:line-numbers}
