+++ apps/docs/src/guide/scripts/custom-script-call.md
@@ -2,4 +2,4 @@
 
 Akin to Contracts, we can configure the [call parameters](../contracts/call-parameters.md) and [transaction parameters](../transactions/transaction-parameters.md) for Scripts, as well as retrieve the entire transaction request or transaction ID prior to submission.
 
-<<< @/../../docs-snippets/src/guide/scripts/script-with-configurable.test.ts#preparing-scripts{ts:line-numbers}
+<<< @/../../docs-snippets2/src/scripts/script-with-configurable.ts#preparing-scripts{ts:line-numbers}
