+++ apps/docs/src/guide/testing/launching-a-test-node.md
@@ -10,7 +10,7 @@ For usage information for `launchTestNode` including it's inputs, outputs and op
 
 We support [explicit resource management](https://www.typescriptlang.org/docs/handbook/variable-declarations.html#using-declarations), introduced in TypeScript 5.2, which automatically calls a `cleanup` function after a variable instantiated with the `using` keyword goes out of block scope:
 
-<<< @/../../docs-snippets/src/guide/testing/launching-a-test-node.test.ts#automatic-cleanup{ts:line-numbers}
+<<< @/../../docs-snippets2/src/testing/launching-a-test-node.ts#automatic-cleanup{ts:line-numbers}
 
 ### Configuring Typescript
 
@@ -35,4 +35,4 @@ If you don't want, or can't use [explicit resource management](https://www.types
 
 In this case, remember you must call `.cleanup()` to dispose of the node.
 
-<<< @/../../docs-snippets/src/guide/testing/launching-a-test-node.test.ts#manual-cleanup{ts:line-numbers}
+<<< @/../../docs-snippets2/src/testing/launching-a-test-node.ts#manual-cleanup{ts:line-numbers}
