+++ apps/docs/src/guide/utilities/unit-conversion.md
@@ -10,17 +10,17 @@ Below we will go over some common use cases for unit conversion.
 
 Using our `BN` class we can instantiate these numbers.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#instantiation-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#instantiation-1{ts:line-numbers}
 
 Or using our `bn` utility function.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#instantiation-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#instantiation-2{ts:line-numbers}
 
 ## Contract calls
 
 Generally, we will need to convert `u64` and `u256` numbers to a `BN` object when passing them to a Sway program from JavaScript. More information on this can be found [here](../types/numbers.md).
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#contract-calls-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#contract-calls-1{ts:line-numbers}
 
 > Note: If a contract call returns a number that is too large to be represented as a JavaScript number, you can convert it to a string using the `toString` method instead of `toNumber`.
 
@@ -28,19 +28,19 @@ Generally, we will need to convert `u64` and `u256` numbers to a `BN` object whe
 
 Parsing string-represented numbers (from user input) has never been easier, than using the `parseUnits` function.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#parse-units-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#parse-units-1{ts:line-numbers}
 
 We can parse large numbers.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#parse-units-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#parse-units-2{ts:line-numbers}
 
 Or numbers formatted for human readability.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#parse-units-3{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#parse-units-3{ts:line-numbers}
 
 We can also parse numbers in other units of measure.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#parse-units-4{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#parse-units-4{ts:line-numbers}
 
 ## Formatting
 
@@ -48,26 +48,33 @@ We can format common units of measure using the `format` function.
 
 In the following example, we format a BigNumber representation of one Gwei, into units for the Fuel network (with 3 decimal place precision).
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#format-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#format-1{ts:line-numbers}
 
 We can also format numbers in other units of measure by specifying the `units` variable.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#format-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#format-2{ts:line-numbers}
 
 A `precision` variable will allow for the formatting of numbers with a specific number of decimal places.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#format-3{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#format-3{ts:line-numbers}
 
 ### Format units
 
 The `formatUnits` function is a lesser alternative to the `format` function, as it will maintain the same precision as the input value.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#format-units-1{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#format-units-1{ts:line-numbers}
 
 We can also format numbers in other units of measure by specifying the `units` variable.
 
-<<< @/../../docs-snippets/src/guide/utilities/unit-conversion.test.ts#format-units-2{ts:line-numbers}
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#format-units-2{ts:line-numbers}
 
 ## See also
 
 - [Sway Numbers](../types/numbers.md)
+
+
+## Full Example
+
+For the full example of unit conversion see the snippet below:
+
+<<< @/../../docs-snippets2/src/utilities/unit-conversion.ts#full{ts:line-numbers}
