+++ tsconfig.test.json
@@ -4,5 +4,5 @@
     "noEmit": true
   },
   "include": ["**/*.test.ts"],
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules", "apps/docs-snippets2/src/**/*.test.ts"]
 }
